Docker configuration updated
The Docker configuration has been updated, splitting it in two different files to avoid running into problems when generating the JAR file for the application.

Before, we couldn't generate the JAR without the database active, but the whole Docker configuration couldn't run properly if the JAR wasn't there. With this configuration we can avoid that problem.